requirements.txt
pandas
sqlalchemy
psycopg2-binary
minio
openpyxl
pyarrow
dbt-core
dbt-postgres
xlrd>=2.0.1
